Refactor: button style.ts 피드백 반영
