+++ src/App.tsx
@@ -1,4 +1,3 @@
-import React from 'react';
 import { GlobalStyle } from './components/globalStyle/globalStyle';
 import Header from './components/header/Header';
 import Main from './components/main/Main';
