+++ src/interfaces/Todo.ts
@@ -1,0 +1,8 @@
+interface Todo {
+  id: string;
+  title: string;
+  content: string;
+  isDone: boolean;
+}
+
+export default Todo
